Move sinks.py into ingestion
