+++ packages/op-datasets/src/op_datasets/etl/ingestion/sinks.py
@@ -2,7 +2,6 @@ import os
 from dataclasses import dataclass
 
 import polars as pl
-import pyarrow as pa
 from op_coreutils import duckdb as utilsduckdb
 from op_coreutils.clickhouse.client import insert_arrow, run_oplabs_query
 from op_coreutils.logger import structlog
@@ -10,82 +9,15 @@ from op_coreutils.storage.gcs_parquet import (
     gcs_upload_parquet,
     local_upload_parquet,
 )
-from op_coreutils.storage.paths import Marker, PartitionedOutput, SinkMarkerPath
-from op_coreutils.time import now
-from overrides import EnforceOverrides
+from op_coreutils.storage.paths import PartitionedOutput, SinkMarkerPath
+
+from op_datasets.etl.ingestion.markers import IngestionProcessMarker
 
 log = structlog.get_logger()
 
 
-@dataclass
-class IngestionProcessMarker:
-    process_name: str
-    chain: str
-    marker: Marker
-
-    @property
-    def path(self):
-        return self.marker.marker_path
-
-    def dt_value(self):
-        return min(_.path.dt_value() for _ in self.marker.outputs)
-
-    def to_clickhouse_pyarrow_table(self) -> pa.Table:
-        row: dict = self.marker.to_clickhouse_row()
-        row["process_name"] = self.process_name
-        row["chain"] = self.chain
-        row["dt"] = self.dt_value()
-
-        marker_schema = self.marker.clickhouse_schema()
-        schema = pa.schema(
-            [
-                marker_schema.field("marker_path"),
-                marker_schema.field("total_rows"),
-                marker_schema.field("outputs.full_path"),
-                marker_schema.field("outputs.partition_cols"),
-                marker_schema.field("outputs.row_count"),
-                pa.field("process_name", pa.string()),
-                pa.field("chain", pa.string()),
-                pa.field("dt", pa.string()),
-            ]
-        )
-
-        return pa.Table.from_pylist(
-            [row],
-            schema=schema,
-        )
-
-    def to_duckdb_pyarrow_table(self) -> pa.Table:
-        row: dict = self.marker.to_duckdb_row()
-        row["updated_at"] = now()
-        row["process_name"] = self.process_name
-        row["chain"] = self.chain
-        row["dt"] = self.dt_value()
-
-        marker_schema = self.marker.duckdb_schema()
-
-        # The order of the columns must match the order as declared in the CREATE TABLE
-        # statement for the table.
-        schema = pa.schema(
-            [
-                pa.field("updated_at", pa.timestamp(unit="us", tz=None)),
-                marker_schema.field("marker_path"),
-                marker_schema.field("total_rows"),
-                marker_schema.field("outputs"),
-                pa.field("process_name", pa.string()),
-                pa.field("chain", pa.string()),
-                pa.field("dt", pa.string()),
-            ]
-        )
-
-        return pa.Table.from_pylist(
-            [row],
-            schema=schema,
-        )
-
-
 @dataclass(kw_only=True)
-class DataSink(EnforceOverrides):
+class DataSink:
     sink_spec: str
 
     def local_full_path(self, path: str):
